+++ medusa/server/api/v2/system.py
@@ -2,13 +2,22 @@
 """Request handler for statistics."""
 from __future__ import unicode_literals
 
+import logging
+import os
+import time
+
 from medusa import app, ui
+from medusa import helpers
+from medusa.logger.adapters.style import BraceAdapter
 from medusa.server.api.v2.base import BaseRequestHandler
 from medusa.system.restart import Restart
 from medusa.system.shutdown import Shutdown
 from medusa.updater.version_checker import CheckVersion
 
 from tornado.escape import json_decode
+
+log = BraceAdapter(logging.getLogger(__name__))
+log.logger.addHandler(logging.NullHandler())
 
 
 class SystemHandler(BaseRequestHandler):
@@ -50,6 +59,7 @@ class SystemHandler(BaseRequestHandler):
                         return self._created()
                     else:
                         return self._bad_request('Update failed')
+                else:
                     return self._bad_request('Backup failed')
             else:
                 ui.notifications.message('Already on branch: ', data['branch'])
@@ -85,6 +95,12 @@ class SystemHandler(BaseRequestHandler):
                 return self._created()
             else:
                 return self._bad_request('Failed starting download handler')
+
+        if data['type'] == 'BACKUPTOZIP':
+            return self._backup_to_zip(data.get('backupDir'))
+
+        if data['type'] == 'RESTOREFROMZIP':
+            return self._restore_from_zip(data.get('backupFile'))
 
         return self._bad_request('Invalid operation')
 
@@ -126,3 +142,63 @@ class SystemHandler(BaseRequestHandler):
             ), 'Check logs for more information.')
 
         return False
+
+    def _backup_to_zip(self, backup_dir):
+        """Create a backup and save to zip."""
+        final_result = ''
+
+        if backup_dir:
+            source = [
+                os.path.join(app.DATA_DIR, app.APPLICATION_DB), app.CONFIG_FILE
+            ]
+
+            if app.BACKUP_CACHE_DB:
+                source += [
+                    os.path.join(app.DATA_DIR, app.FAILED_DB),
+                    os.path.join(app.DATA_DIR, app.CACHE_DB),
+                    os.path.join(app.DATA_DIR, app.RECOMMENDED_DB)
+                ]
+            target = os.path.join(backup_dir, 'medusa-{date}.zip'.format(date=time.strftime('%Y%m%d%H%M%S')))
+            log.info(u'Starting backup to location: {location} ', {'location': target})
+
+            if app.BACKUP_CACHE_FILES:
+                for (path, dirs, files) in os.walk(app.CACHE_DIR, topdown=True):
+                    for dirname in dirs:
+                        if path == app.CACHE_DIR and dirname not in ['images']:
+                            dirs.remove(dirname)
+                    for filename in files:
+                        source.append(os.path.join(path, filename))
+
+            if helpers.backup_config_zip(source, target, app.DATA_DIR):
+                final_result += 'Successful backup to {location}'.format(location=target)
+            else:
+                final_result += 'Backup FAILED'
+        else:
+            final_result += 'You need to choose a folder to save your backup to!'
+
+        final_result += '<br>\n'
+
+        log.info(u'Finished backup to location: {location} ', {'location': target})
+        return self._ok(data={'result': final_result})
+
+    def _restore_from_zip(self, backup_file):
+        """Restore from zipped backup."""
+        final_result = ''
+
+        if backup_file:
+            source = backup_file
+            target_dir = os.path.join(app.DATA_DIR, 'restore')
+            log.info(u'Restoring backup from location: {location} ', {'location': backup_file})
+
+            if helpers.restore_config_zip(source, target_dir):
+                final_result += 'Successfully extracted restore files to {location}'.format(location=target_dir)
+                final_result += '<br>Restart Medusa to complete the restore.'
+            else:
+                final_result += 'Restore FAILED'
+        else:
+            final_result += 'You need to select a backup file to restore!'
+
+        final_result += '<br>\n'
+
+        log.info(u'Finished restore from location: {location}', {'location': backup_file})
+        return self._ok(data={'result': final_result})