+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.5.29 (11-04-2022)
+
+#### New Features
+- Support for Plex metadata (.plexmatch) ([10466](https://github.com/pymedusa/Medusa/pull/10466)) 
+
+#### Improvements
+- Make the cache db and cache files optional for inclusion in the backup ([10475](https://github.com/pymedusa/Medusa/pull/10475))
+
+#### Fixes
+- Fix joining segments in log for failed episodes ([10472](https://github.com/pymedusa/Medusa/pull/10472))
+
 ## 0.5.28 (01-04-2022)
 
 #### Improvements